SAXParser: write: Clear fragments after prepending to current source
Currently, when a chunk ends with an incomplete UTF-8 sequence,
the bytes are stored in self.fragment and prepended to the next
chunk.
However, after prepending the bytes, self.fragments isn't cleared
in all cases. This leads to a corruption of all future chunks, as
the same incomplete UTF-8 sequence is prepended to every chunk
until there's another incomplete UTF-8 sequence at the end of a
chunk.

Therefore, self.fragment is now cleared immediately after
prepending the bytes to the next chunk.